Set macho::symbol_table enums to match their field-types

+++ src/mach-o/headers/symbol_table.h
@@ -36,14 +36,14 @@ namespace macho {
     };
 
     namespace symbol_table {
-        enum class flags : int {
+        enum class flags : uint8_t {
             stab = 0xe0,
             private_external = 0x10,
             type = 0x0e,
             external = 0x01
         };
 
-        enum class type : int {
+        enum class type : uint8_t {
             undefined,
             absolute = 0x2,
             section = 0xe,
@@ -51,7 +51,7 @@ namespace macho {
             indirect = 0xa
         };
 
-        enum description : int {
+        enum description : uint16_t {
             no_dead_strip = 0x0020,
             weak_reference = 0x0040,
             weak_definition = 0x0080,
